+++ src/OpenTelemetry/AssemblyInfo.cs
@@ -22,5 +22,7 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Benchmarks" + AssemblyInfo.PublicKey)]
 
 // TODO: Much of the metrics SDK is currently internal. These should be removed once the public API surface area for metrics is defined.
+[assembly: InternalsVisibleTo("OpenTelemetry.Exporter.Jaeger" + AssemblyInfo.PublicKey)]
 [assembly: InternalsVisibleTo("OpenTelemetry.Exporter.OpenTelemetryProtocol" + AssemblyInfo.PublicKey)]
 [assembly: InternalsVisibleTo("OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests" + AssemblyInfo.PublicKey)]
+[assembly: InternalsVisibleTo("OpenTelemetry.Exporter.Zipkin" + AssemblyInfo.PublicKey)]