+++ src/OpenTelemetry.Exporter.Prometheus/PrometheusExporter.cs
@@ -15,6 +15,7 @@
 // </copyright>
 
 using System;
+using System.Collections.Generic;
 using OpenTelemetry.Metrics;
 
 namespace OpenTelemetry.Exporter
@@ -22,10 +23,11 @@ namespace OpenTelemetry.Exporter
     /// <summary>
     /// Exporter of OpenTelemetry metrics to Prometheus.
     /// </summary>
-    public class PrometheusExporter : BaseExporter<MetricItem>
+    [AggregationTemporality(AggregationTemporality.Cumulative)]
+    public class PrometheusExporter : BaseExporter<Metric>
     {
         internal readonly PrometheusExporterOptions Options;
-        internal Batch<MetricItem> Batch;
+        internal Batch<Metric> Metrics;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="PrometheusExporter"/> class.
@@ -36,11 +38,11 @@ namespace OpenTelemetry.Exporter
             this.Options = options;
         }
 
-        internal Action MakePullRequest { get; set; }
+        internal Action CollectMetric { get; set; }
 
-        public override ExportResult Export(in Batch<MetricItem> batch)
+        public override ExportResult Export(in Batch<Metric> metrics)
         {
-            this.Batch = batch;
+            this.Metrics = metrics;
             return ExportResult.Success;
         }
     }