+++ src/OpenTelemetry/CHANGELOG.md
@@ -5,6 +5,14 @@
 * `BatchExportProcessor.OnShutdown` will now log the count of dropped telemetry items.
   ([#2331](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2331))
 
+## 1.2.0-alpha3
+
+Released 2021-Sep-13
+
+* Metrics perf improvements, bug fixes.
+  Replace MetricProcessor with MetricReader.
+  ([#2306](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2306))
+
 * Add `BatchExportActivityProcessorOptions` which supports field value overriding
   using `OTEL_BSP_SCHEDULE_DELAY`, `OTEL_BSP_EXPORT_TIMEOUT`,
   `OTEL_BSP_MAX_QUEUE_SIZE`, `OTEL_BSP_MAX_EXPORT_BATCH_SIZE`